+++ packages/plugin-truffle/src/prepare-upgrade.ts
@@ -3,7 +3,7 @@ import {
   isBeacon,
   isBeaconProxy,
   isTransparentOrUUPSProxy,
-  PrepareUpgradeUnsupportedError,
+  PrepareUpgradeRequiresKindError,
 } from '@openzeppelin/upgrades-core';
 import {
   ContractClass,
@@ -14,26 +14,30 @@ import {
   wrapProvider,
   deployBeaconImpl,
   PrepareUpgradeOptions,
+  deployUpgradeableImpl,
 } from './utils';
 
 export async function prepareUpgrade(
-  proxyOrBeacon: ContractAddressOrInstance,
+  referenceAddressOrContract: ContractAddressOrInstance,
   Contract: ContractClass,
   opts: PrepareUpgradeOptions = {},
 ): Promise<string> {
-  const proxyOrBeaconAddress = getContractAddress(proxyOrBeacon);
+  const referenceAddress = getContractAddress(referenceAddressOrContract);
   const { deployer } = withDefaults(opts);
   const provider = wrapProvider(deployer.provider);
   let deployedImpl;
-  if (await isTransparentOrUUPSProxy(provider, proxyOrBeaconAddress)) {
-    deployedImpl = await deployProxyImpl(Contract, opts, proxyOrBeaconAddress);
-  } else if (await isBeaconProxy(provider, proxyOrBeaconAddress)) {
-    const beaconAddress = await getBeaconAddress(provider, proxyOrBeaconAddress);
+  if (await isTransparentOrUUPSProxy(provider, referenceAddress)) {
+    deployedImpl = await deployProxyImpl(Contract, opts, referenceAddress);
+  } else if (await isBeaconProxy(provider, referenceAddress)) {
+    const beaconAddress = await getBeaconAddress(provider, referenceAddress);
     deployedImpl = await deployBeaconImpl(Contract, opts, beaconAddress);
-  } else if (await isBeacon(provider, proxyOrBeaconAddress)) {
-    deployedImpl = await deployBeaconImpl(Contract, opts, proxyOrBeaconAddress);
+  } else if (await isBeacon(provider, referenceAddress)) {
+    deployedImpl = await deployBeaconImpl(Contract, opts, referenceAddress);
   } else {
-    throw new PrepareUpgradeUnsupportedError(proxyOrBeaconAddress);
+    if (opts.kind === undefined) {
+      throw new PrepareUpgradeRequiresKindError();
+    }
+    deployedImpl = await deployUpgradeableImpl(Contract, opts, referenceAddress);
   }
   return deployedImpl.impl;
 }