+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 1.25.0 (2023-04-26)
+
+- Add support for Arbitrum to manifest file names. ([#770](https://github.com/OpenZeppelin/openzeppelin-upgrades/pull/770))
+- Add support for Sepolia testnet to manifest file names. ([#766](https://github.com/OpenZeppelin/openzeppelin-upgrades/pull/766))
+- Support `prepareUpgrade` from an implementation address. ([#777](https://github.com/OpenZeppelin/openzeppelin-upgrades/issues/777))
+
 ## 1.24.1 (2023-03-02)
 
 - Remove test contracts from source code verification. ([#751](https://github.com/OpenZeppelin/openzeppelin-upgrades/pull/751))