+++ packages/plugin-hardhat/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 1.23.1 (2023-04-26)
+
+- Enable `verify` to fall back to hardhat-etherscan if proxy bytecode does not match. ([#752](https://github.com/OpenZeppelin/openzeppelin-upgrades/issues/752))
+
+## 1.23.0 (2023-04-26)
+
+- Support `prepareUpgrade` from an implementation address. ([#777](https://github.com/OpenZeppelin/openzeppelin-upgrades/issues/777))
+
 ## 1.22.1 (2023-01-18)
 
 - Handle getLogs error for Blockscout explorer. ([#706](https://github.com/OpenZeppelin/openzeppelin-upgrades/pull/706))